requirements.txt
fastmcp==2.11.3
mcp-use==1.3.10
streamlit==1.49.1
openai==1.102.0
pandas==2.3.2
chardet==5.2.0
langchain-openai==0.3.32
python-dotenv==1.1.1
mcp==1.13.1
nest_asyncio
